Fix breaking changes from OSS Airflow
A few months back, the SnowflakeOperator was changed (...by me)
to use `get_db_hook()` instead of `get_hook` in order to make it
inheritable by BaseSQLOperator. This allowed SnowflakeCheckOperators
to inherit from BaseSQLOperator and simplify backend connections. This
change broke the extractor in OpenLineage, and this PR aims to fix it.

Signed-off-by: Benji Lampel <denimalpaca@gmail.com>
Signed-off-by: Benji Lampel <benjamin@astronomer.io>