[INTEGRATION][AIRFLOW] Fix Breaking SnowflakeOperator Changes from OSS Airflow (#507)
* Fix breaking changes from OSS Airflow

A few months back, the SnowflakeOperator was changed (...by me)
to use `get_db_hook()` instead of `get_hook` in order to make it
inheritable by BaseSQLOperator. This allowed SnowflakeCheckOperators
to inherit from BaseSQLOperator and simplify backend connections. This
change broke the extractor in OpenLineage, and this PR aims to fix it.

Signed-off-by: Benji Lampel <denimalpaca@gmail.com>
Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* Allow backward compatibility with old version of SnowflakeOperator

Add if statements to check if the operator is using get_hook() or
get_db_hook(), allowing for backward compatibility.

Signed-off-by: Benji Lampel <denimalpaca@gmail.com>
Signed-off-by: Benji Lampel <benjamin@astronomer.io>

Co-authored-by: Maciej Obuchowski <maciej.obuchowski@getindata.com>