dagster: update unit test to handle updated EventLogEntry
In Dagster 0.14.3, `message` was removed from EventLogEntry. This is Dagster's internal change that should not affect the functionality of the OpenLineage sensor, but the unit test helper function that creates a mock EventLogEntry needs an update to handle this differently based on Dagster version.

Signed-off-by: Dalin Kim <dalinkim@northwesternmutual.com>

dagster: remove unnecessary dict look up

Signed-off-by: Dalin Kim <dalinkim@northwesternmutual.com>