Support Azure Databricks Credential Passthrough
Closes #576 by attempting to extract HadoopFsRelation paths only if there is an exception.
The Azure Databricks Credential Passthrough attempts to use a custom auth when connecting to ABFSS.
As a result, the SparkListener is not running in a context that the custom auth can get the current user's identity, resulting in an exception (com.databricks.backend.daemon.data.client.adl.AzureCredentialNotFoundException).

The defining based off of relation URI paths is similar to InsertIntoHadoopFsRelationVisitor.

Signed-off-by: Will Johnson <will@willj.co>