Fix action.py to pass test_dataset_from_sql_source_v3_api
In the V3 API, the Validator object contains a Batch object
that contains the table, schema, etc data that the lineage
action needs. However, the schema_name property won't always
exist (BigQuery, e.g.), and so a try/except clause is added
when fetching that value from the dictionary. This lack of
schema is already handled in the following function,
_get_sql_table(), so this behavior was already known and
accounted for, but too late now with the V3 API changes.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>