Major updates on extractors
Updated and bug-fixed new extractor(s), and added
more in-depth parsing of check results in util
functions. Additionally, changed the extract()
method in the base extractor to get the task instance
from the on_complete version, as the task instance is
needed by the SQL check extractors to get their XCOM
results.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>