Refactor Check Extractors after latest rebase
The latest rebase included a commit that restructured
the SQL based extractors by adding a SqlExtractor that
these all inherit from. A chagne was made in the SqlExtractor
to allow input and output datasets to build specific input and
output facets; these are implemented and used in the check operators.
Additionally, the check operators are refactored to make use of
common code from the SqlExtractor and are therefore much simpler.
The provider-based extractors use multiple inhertance to take
advantage of the _get_db_specific_run_facets() methods. However,
there is a drawback to this approach: if someone is using a
SQLCheckOperator utilizing hook_params() to call a provider's
hook instead of the provider version of the operator, then the
SqlCheckExtractor will run instead of the provider-specific one,
which will not call _get_db_specific_run_facets().

Signed-off-by: Benji Lampel <benjamin@astronomer.io>