Overhauled SqlCheckExtractors to use Dynamic Inheritance
SqlCheckExtractors all now inherit from abstract class
BaseSqlCheckExtractor, which dynamically inherits from a
non-abstract extractor class at runtime. This is done to
minimize the amount of new operators and testing needed per
provider. When the Extractor class is initialized, it adds
the SqlCheckExtractors as abstract classes inheriting from
SqlExtractor, then when Manager calls _get_extractor,
the Extractor.instantiate_abstract_extractors() method is called.
This method looks up relevant extractor/conn_type pairs in a new
dict at the top of the file, and matches the parent extractor
instance for the SqlCheckOperators with the task by conn_type.
This ensures that when a SqlCheckOperator is run with a given
connection type, the appropriate impementation of extract() or
extract_on_complete() is run at the provider level.

Currently, BigQuery does not work, as the SqlCheckOperators do not
emit a job id. This is an issue that should be fixed in OSS Airflow.

Testing is also currently broken.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>