Create 'scala-fixtures' module and refactor some tests
This adds the capability to write Spark jobs using the Scala API, and
have them run as part of the integration tests inside 'app'.

The reason for this, is that there was a change in Scala 2.13 that
caused the `spark-shell -i file.scala' command to misbehave.

This overcomes that, by removing the 'rdd_union.scala' script, and
replacing it with this JAR.

Additionally, created the SparkContainerProperties class in order to
collect the system properties needed to run the integration tests. This
provides flexibility by relying on configuration instead of hard coding.

Removed some of the redundant test tags, like 'beforeShadowJarTest' and
'nonParallelTest'.

Fixed a newly revealed bug in the SparkSQLExecutionContextTest, when
trying to serialise the mock logical plan. Altered the test and the
SparkAgentExtension class. The SparkAgentExtension class now disables
the 'spark_unknown' facet.

Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>