spark(OpenLineageContext): Add direct accessors for the LogicalPlan and SparkPlan objects
These methods surface the LogicalPlan and SparkPlan objects on the underlying QueryExecution object.
The codebase makes extensive use of the following pattern:

context.getQueryExecution().get().<method>

This violates the Law of Demeter as we are reaching through the QueryExecution object via the OpenLineageContext object.
Hence, the QueryExecution accessor is deprecated and callers should prefer to use one of the direct methods.

By encapsulating the QueryExecution object and exposing accessors for the plans, we can also do other things like plan splicing.

Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>