doc: Update contributing/development docs (#4010)
* doc: Update PR creation docs

There is a bot configured to add relevant tags automatically to the PR,
there is no need to add them manually, but they should be reviewed
anyway.

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Remove section about branching groups

This convention does not make much sense in the fork and pull model.
Within the repo itself it is not utilized anyway. Much more often we
have branches without the group or with the component name as group
(e.g. `spark/new-feature`)

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Update pre-commit section

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Remove mention of CI script

It seems like this script no longer has any use, and it might be
confusing for new contributors to see it mentioned here.
The CI runs automatically for PRs from forks.

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Move license header doc to CONTRIBUTING.md

Move details about license header when making contributions to
CONTRIBUTING.md. Having it all in one place makes the doc easier
to follow.

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Update docs on license header

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Link to the dev docs for new contributors

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Remove links to env variables

This is an unnecessary duplication, Python/Java client docs are already
linked above, where it is easy to find env variables or any other config
details.

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Add mention of Java version of SQL parser

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Add base Java setup docs

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Nest Spark integration docs inside Java

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Remove duplicated license header mention

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Update spark integration setup doc

Signed-off-by: kchledowski <github@chledowski.com>

* doc: Update mention of C compiler for Rust

Signed-off-by: kchledowski <github@chledowski.com>

---------

Signed-off-by: kchledowski <github@chledowski.com>